lib removal
