+++ frontend/src/metabase/visualizations/lib/formatting/value.tsx
@@ -5,10 +5,6 @@ import ReactMarkdown from "react-markdown";
 
 import { ExternalLink } from "metabase/common/components/ExternalLink";
 import CS from "metabase/css/core/index.css";
-import {
-  clickBehaviorIsValid,
-  getDataFromClicked,
-} from "metabase/parameters/utils/click-behavior";
 import { NULL_DISPLAY_VALUE } from "metabase/utils/constants";
 import { formatNumber } from "metabase/utils/formatting/numbers";
 import { removeNewLines } from "metabase/utils/formatting/strings";
@@ -25,6 +21,7 @@ import {
 } from "metabase-lib/v1/types/utils/isa";
 import type { ColumnSettings, DatasetColumn } from "metabase-types/api";
 
+import { clickBehaviorIsValid, getDataFromClicked } from "./click-data";
 import { formatDateTimeWithUnit, formatRange } from "./date";
 import { formatEmail } from "./email";
 import { formatCoordinate } from "./geography";
